chore: clean up prb math

+++ test/Fenix.t.sol
@@ -88,7 +88,7 @@ contract FenixTest is Test {
         uint256 bonus = 2.77069 * 1e18;
 
         uint256 timestamp = block.timestamp;
-        Stake memory stake1 = Stake(uint40(timestamp), 1, 100, base, bonus, base + bonus);
+        Stake memory stake1 = Stake(timestamp, 1, 100, base, bonus, base + bonus);
 
         vm.warp(timestamp + (86400 * 0));
         uint256 penalty0 = fenix.calculateEarlyPenalty(stake1);
@@ -117,7 +117,7 @@ contract FenixTest is Test {
         uint256 baseTerm = 100;
 
         uint256 timestamp = block.timestamp;
-        Stake memory stake1 = Stake(uint40(timestamp), 1, baseTerm, base, bonus, base + bonus);
+        Stake memory stake1 = Stake(timestamp, 1, baseTerm, base, bonus, base + bonus);
 
         vm.warp(timestamp + (86400 * baseTerm));
         uint256 penalty0 = fenix.calculateLatePenalty(stake1);
